Move versioned external IP types from common to sled-agent-types
