+++ clients/sled-agent-client/src/lib.rs
@@ -57,9 +57,9 @@ progenitor::generate_api!(
         DiskVariant = omicron_common::disk::DiskVariant,
         Epoch = trust_quorum_types::types::Epoch,
         ExternalIpGatewayMap = omicron_common::api::internal::shared::ExternalIpGatewayMap,
-        ExternalIpConfig = omicron_common::api::internal::shared::ExternalIpConfig,
-        ExternalIpv4Config = omicron_common::api::internal::shared::ExternalIpv4Config,
-        ExternalIpv6Config = omicron_common::api::internal::shared::ExternalIpv6Config,
+        ExternalIpConfig = sled_agent_types_versions::latest::instance::ExternalIpConfig,
+        ExternalIpv4Config = sled_agent_types_versions::latest::instance::ExternalIpv4Config,
+        ExternalIpv6Config = sled_agent_types_versions::latest::instance::ExternalIpv6Config,
         Generation = omicron_common::api::external::Generation,
         Hostname = omicron_common::api::external::Hostname,
         ImportExportPolicy = sled_agent_types_versions::latest::early_networking::ImportExportPolicy,
@@ -102,7 +102,7 @@ progenitor::generate_api!(
         RouterVersion = omicron_common::api::internal::shared::RouterVersion,
         Sha3_256Digest = sled_agent_types_versions::latest::rot::Sha3_256Digest,
         SledRole = sled_agent_types_versions::latest::inventory::SledRole,
-        SourceNatConfigGeneric = omicron_common::api::internal::shared::SourceNatConfigGeneric,
+        SourceNatConfigGeneric = sled_agent_types_versions::latest::inventory::SourceNatConfigGeneric,
         SwitchSlot = sled_agent_types_versions::latest::early_networking::SwitchSlot,
         SystemNetworkingConfig = sled_agent_types_versions::latest::system_networking::SystemNetworkingConfig,
         Threshold = trust_quorum_types::types::Threshold,
