+++ illumos-utils/src/opte/port_manager.rs
@@ -20,10 +20,7 @@ use macaddr::MacAddr6;
 use omicron_common::address::IPV4_MULTICAST_RANGE;
 use omicron_common::address::IPV6_MULTICAST_RANGE;
 use omicron_common::api::external;
-use omicron_common::api::internal::shared::ExternalIpConfig;
 use omicron_common::api::internal::shared::ExternalIpGatewayMap;
-use omicron_common::api::internal::shared::ExternalIpv4Config;
-use omicron_common::api::internal::shared::ExternalIpv6Config;
 use omicron_common::api::internal::shared::InternetGatewayRouterTarget;
 use omicron_common::api::internal::shared::NetworkInterface;
 use omicron_common::api::internal::shared::NetworkInterfaceKind;
@@ -61,6 +58,9 @@ use oxide_vpc::api::VpcCfg;
 use oxnet::IpNet;
 use oxnet::Ipv4Net;
 use oxnet::Ipv6Net;
+use sled_agent_types::instance::ExternalIpConfig;
+use sled_agent_types::instance::ExternalIpv4Config;
+use sled_agent_types::instance::ExternalIpv6Config;
 use slog::Logger;
 use slog::debug;
 use slog::error;
@@ -1232,9 +1232,6 @@ mod tests {
     use crate::opte::Handle;
     use macaddr::MacAddr6;
     use omicron_common::api::external::{MacAddr, Vni};
-    use omicron_common::api::internal::shared::ExternalIpConfig;
-    use omicron_common::api::internal::shared::ExternalIpv4Config;
-    use omicron_common::api::internal::shared::ExternalIpv6Config;
     use omicron_common::api::internal::shared::InternetGatewayRouterTarget;
     use omicron_common::api::internal::shared::NetworkInterface;
     use omicron_common::api::internal::shared::NetworkInterfaceKind;
@@ -1245,8 +1242,6 @@ mod tests {
     use omicron_common::api::internal::shared::ResolvedVpcRouteSet;
     use omicron_common::api::internal::shared::RouterTarget;
     use omicron_common::api::internal::shared::RouterVersion;
-    use omicron_common::api::internal::shared::SourceNatConfigV4;
-    use omicron_common::api::internal::shared::SourceNatConfigV6;
     use omicron_test_utils::dev::test_setup_log;
     use oxide_vpc::api::DhcpCfg;
     use oxide_vpc::api::IpCfg;
@@ -1255,6 +1250,11 @@ mod tests {
     use oxnet::IpNet;
     use oxnet::Ipv4Net;
     use oxnet::Ipv6Net;
+    use sled_agent_types::instance::ExternalIpConfig;
+    use sled_agent_types::instance::ExternalIpv4Config;
+    use sled_agent_types::instance::ExternalIpv6Config;
+    use sled_agent_types::inventory::SourceNatConfigV4;
+    use sled_agent_types::inventory::SourceNatConfigV6;
     use std::collections::HashSet;
     use std::net::Ipv4Addr;
     use std::net::Ipv6Addr;
