+++ nexus/db-model/src/deployment.rs
@@ -70,6 +70,7 @@ use omicron_uuid_kinds::{
     PhysicalDiskKind, SledKind, SledUuid, ZpoolKind, ZpoolUuid,
 };
 use sled_agent_types::inventory::OmicronZoneDataset;
+use sled_agent_types::inventory::SourceNatConfigGeneric;
 use sled_hardware_types::BaseboardId;
 use std::net::{IpAddr, SocketAddrV6};
 use std::sync::Arc;
@@ -923,7 +924,7 @@ impl BpOmicronZone {
                     self.snat_last_port,
                 ) {
                     (Some(ip), Some(first_port), Some(last_port)) => {
-                        nexus_types::inventory::SourceNatConfigGeneric::new(
+                        SourceNatConfigGeneric::new(
                             ip.ip(),
                             *first_port,
                             *last_port,
