+++ nexus/db-queries/src/db/datastore/external_ip.rs
@@ -1351,7 +1351,6 @@ mod tests {
     use nexus_types::external_api::instance::PrivateIpStackCreate;
     use nexus_types::external_api::vpc;
     use nexus_types::identity::Resource;
-    use nexus_types::inventory::SourceNatConfigGeneric;
     use omicron_common::address::IpRange;
     use omicron_common::address::Ipv4Range;
     use omicron_common::address::NUM_SOURCE_NAT_PORTS;
@@ -1360,6 +1359,7 @@ mod tests {
     };
     use omicron_test_utils::dev;
     use omicron_uuid_kinds::ExternalIpUuid;
+    use sled_agent_types::inventory::SourceNatConfigGeneric;
     use std::collections::BTreeSet;
     use std::net::Ipv4Addr;
