+++ nexus/types/src/deployment/network_resources.rs
@@ -9,13 +9,13 @@ use iddqd::TriHashMap;
 use iddqd::tri_upcast;
 use omicron_common::api::external::IpVersion;
 use omicron_common::api::external::MacAddr;
-use omicron_common::api::internal::shared::SourceNatConfigGeneric;
 use omicron_uuid_kinds::ExternalIpUuid;
 use omicron_uuid_kinds::OmicronZoneUuid;
 use omicron_uuid_kinds::VnicUuid;
 use schemars::JsonSchema;
 use serde::Deserialize;
 use serde::Serialize;
+use sled_agent_types::inventory::SourceNatConfigGeneric;
 use std::net::IpAddr;
 use std::net::SocketAddr;
 use thiserror::Error;
