+++ sled-agent/src/instance.rs
@@ -29,8 +29,7 @@ use illumos_utils::zone::PROPOLIS_ZONE_PREFIX;
 use illumos_utils::zpool::ZpoolOrRamdisk;
 use omicron_common::api::internal::nexus::{SledVmmState, VmmRuntimeState};
 use omicron_common::api::internal::shared::{
-    DelegatedZvol, ExternalIpConfig, NetworkInterface, ResolvedVpcFirewallRule,
-    SledIdentifiers,
+    DelegatedZvol, NetworkInterface, ResolvedVpcFirewallRule, SledIdentifiers,
 };
 use omicron_common::backoff;
 use omicron_common::backoff::BackoffError;
