+++ sled-agent/tests/integration_tests/early_network.rs
@@ -6,7 +6,6 @@
 
 use bootstore::schemes::v0 as bootstore;
 use iddqd::IdOrdMap;
-use nexus_types::inventory::SourceNatConfigGeneric;
 use omicron_common::api::external::Vni;
 use omicron_test_utils::dev::test_setup_log;
 use sled_agent_types::early_networking::{
@@ -15,6 +14,7 @@ use sled_agent_types::early_networking::{
     PortSpeed, RackNetworkConfig, RouterLifetimeConfig, RouterPeerType,
     SwitchSlot, UplinkAddress, UplinkAddressConfig,
 };
+use sled_agent_types::inventory::SourceNatConfigGeneric;
 use sled_agent_types::system_networking::{
     ServiceZoneNatEntries, ServiceZoneNatEntry, ServiceZoneNatKind,
     SystemNetworkingConfig,
