+++ sled-agent/types/versions/Cargo.toml
@@ -17,6 +17,7 @@ chrono.workspace = true
 daft.workspace = true
 iddqd.workspace = true
 indent_write.workspace = true
+itertools.workspace = true
 omicron-common.workspace = true
 omicron-ledger.workspace = true
 omicron-passwords.workspace = true
