+++ sled-agent/types/versions/src/add_attached_subnets/instance.rs
@@ -5,7 +5,6 @@ use omicron_common::api::internal::nexus::VmmRuntimeState;
 use omicron_common::api::internal::shared::DelegatedZvol;
 use omicron_common::api::internal::shared::DhcpConfig;
 use omicron_common::api::internal::shared::NetworkInterface;
-use omicron_common::api::internal::shared::external_ip::v1::ExternalIpConfig;
 use omicron_uuid_kinds::InstanceUuid;
 
 use crate::v10::instance::ResolvedVpcFirewallRule;
@@ -17,6 +16,7 @@ use uuid::Uuid;
 use crate::v1::instance::InstanceMetadata;
 use crate::v1::instance::VmmSpec;
 use crate::v7::instance::InstanceMulticastMembership;
+use crate::v11::instance::ExternalIpConfig;
 use crate::v18::attached_subnet::AttachedSubnet;
 
 use crate::v17;
