+++ sled-agent/types/versions/src/add_dual_stack_external_ip_config/instance.rs
@@ -2,19 +2,28 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+use std::net::IpAddr;
+use std::net::Ipv4Addr;
+use std::net::Ipv6Addr;
 use std::net::SocketAddr;
 
+use itertools::Either;
+use itertools::Itertools;
 use omicron_common::api::external;
 use omicron_common::api::external::Hostname;
 use omicron_common::api::internal::nexus::VmmRuntimeState;
 use omicron_common::api::internal::shared::DhcpConfig;
 use omicron_common::api::internal::shared::NetworkInterface;
-use omicron_common::api::internal::shared::external_ip::v1::ExternalIpConfig;
 use omicron_uuid_kinds::InstanceUuid;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
+use super::inventory::SourceNatConfig;
+use super::inventory::SourceNatConfigV4;
+use super::inventory::SourceNatConfigV6;
+use crate::impls::inventory::SourceNatConfigError;
+use crate::v1;
 use crate::v1::instance::InstanceMetadata;
 use crate::v1::instance::VmmSpec;
 use crate::v7::instance::InstanceMulticastMembership;
@@ -91,7 +100,6 @@ impl TryFrom<v10::instance::InstanceSledLocalConfig>
     fn try_from(
         v10: v10::instance::InstanceSledLocalConfig,
     ) -> Result<Self, Self::Error> {
-        // v10.source_nat is already a v1::SourceNatConfig, so we can use it directly
         let external_ips = ExternalIpConfig::try_from_generic(
             Some(v10.source_nat),
             v10.ephemeral_ip,
@@ -113,4 +121,211 @@ impl TryFrom<v10::instance::InstanceSledLocalConfig>
             delegated_zvols: v10.delegated_zvols,
         })
     }
+}
+
+/// IPv4 external IP configuration.
+#[derive(
+    Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize,
+)]
+pub struct ExternalIpv4Config {
+    /// Source NAT configuration, for outbound-only connectivity.
+    pub source_nat: Option<SourceNatConfigV4>,
+    /// An Ephemeral address for in- and outbound connectivity.
+    pub ephemeral_ip: Option<Ipv4Addr>,
+    /// Additional Floating IPs for in- and outbound connectivity.
+    pub floating_ips: Vec<Ipv4Addr>,
+}
+
+/// IPv6 external IP configuration.
+#[derive(
+    Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize,
+)]
+pub struct ExternalIpv6Config {
+    /// Source NAT configuration, for outbound-only connectivity.
+    pub source_nat: Option<SourceNatConfigV6>,
+    /// An Ephemeral address for in- and outbound connectivity.
+    pub ephemeral_ip: Option<Ipv6Addr>,
+    /// Additional Floating IPs for in- and outbound connectivity.
+    pub floating_ips: Vec<Ipv6Addr>,
+}
+
+/// A single- or dual-stack external IP configuration.
+#[derive(Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize)]
+#[serde(rename_all = "snake_case", tag = "type", content = "value")]
+pub enum ExternalIpConfig {
+    /// Single-stack IPv4 external IP configuration.
+    V4(ExternalIpv4Config),
+    /// Single-stack IPv6 external IP configuration.
+    V6(ExternalIpv6Config),
+    /// Both IPv4 and IPv6 external IP configuration.
+    DualStack { v4: ExternalIpv4Config, v6: ExternalIpv6Config },
+}
+
+impl From<ExternalIpv4Config> for ExternalIpConfig {
+    fn from(cfg: ExternalIpv4Config) -> Self {
+        Self::V4(cfg)
+    }
+}
+
+impl From<ExternalIpv6Config> for ExternalIpConfig {
+    fn from(cfg: ExternalIpv6Config) -> Self {
+        Self::V6(cfg)
+    }
+}
+
+// This impl is here because it's only used in the `TryFrom<_>` above.
+impl ExternalIpConfig {
+    /// Attempt to convert from generic IP addressing information.
+    ///
+    /// This is used to convert older API versions which used version-agnostic
+    /// IP address types (`IpAddr`) rather than concrete `Ipv4Addr`/`Ipv6Addr`.
+    ///
+    /// Returns an error if there are no addresses at all, or if there are
+    /// mixed IP versions.
+    fn try_from_generic(
+        source_nat: Option<v1::inventory::SourceNatConfig>,
+        ephemeral_ip: Option<IpAddr>,
+        floating_ips: Vec<IpAddr>,
+    ) -> Result<Self, ExternalIpsError> {
+        let snat_v4;
+        let snat_v6;
+        match source_nat {
+            Some(snat) => match snat.ip {
+                IpAddr::V4(ipv4) => {
+                    snat_v6 = None;
+                    snat_v4 = Some(SourceNatConfig::new(
+                        ipv4,
+                        snat.first_port,
+                        snat.last_port,
+                    )?);
+                }
+                IpAddr::V6(ipv6) => {
+                    snat_v6 = Some(SourceNatConfig::new(
+                        ipv6,
+                        snat.first_port,
+                        snat.last_port,
+                    )?);
+                    snat_v4 = None;
+                }
+            },
+            None => {
+                snat_v4 = None;
+                snat_v6 = None;
+            }
+        };
+
+        let (fips_v4, fips_v6): (Vec<_>, Vec<_>) =
+            floating_ips.into_iter().partition_map(|fip| match fip {
+                IpAddr::V4(v4) => Either::Left(v4),
+                IpAddr::V6(v6) => Either::Right(v6),
+            });
+
+        match (snat_v4, snat_v6, ephemeral_ip) {
+            (None, None, None) => {
+                match (fips_v4.is_empty(), fips_v6.is_empty()) {
+                    (true, true) => Err(ExternalIpsError::NoIps),
+                    (true, false) => Ok(ExternalIpv6Config {
+                        source_nat: None,
+                        ephemeral_ip: None,
+                        floating_ips: fips_v6,
+                    }
+                    .into()),
+                    (false, true) => Ok(ExternalIpv4Config {
+                        source_nat: None,
+                        ephemeral_ip: None,
+                        floating_ips: fips_v4,
+                    }
+                    .into()),
+                    (false, false) => Err(ExternalIpsError::MixedIpVersions),
+                }
+            }
+            (None, None, Some(IpAddr::V4(v4))) => {
+                if !fips_v6.is_empty() {
+                    return Err(ExternalIpsError::MixedIpVersions);
+                }
+                Ok(ExternalIpv4Config {
+                    source_nat: None,
+                    ephemeral_ip: Some(v4),
+                    floating_ips: fips_v4,
+                }
+                .into())
+            }
+            (None, None, Some(IpAddr::V6(v6))) => {
+                if !fips_v4.is_empty() {
+                    return Err(ExternalIpsError::MixedIpVersions);
+                }
+                Ok(ExternalIpv6Config {
+                    source_nat: None,
+                    ephemeral_ip: Some(v6),
+                    floating_ips: fips_v6,
+                }
+                .into())
+            }
+            (None, Some(snat_v6), None) => {
+                if !fips_v4.is_empty() {
+                    return Err(ExternalIpsError::MixedIpVersions);
+                }
+                Ok(ExternalIpv6Config {
+                    source_nat: Some(snat_v6),
+                    ephemeral_ip: None,
+                    floating_ips: fips_v6,
+                }
+                .into())
+            }
+            (None, Some(snat_v6), Some(IpAddr::V6(eip_v6))) => {
+                if !fips_v4.is_empty() {
+                    return Err(ExternalIpsError::MixedIpVersions);
+                }
+                Ok(ExternalIpv6Config {
+                    source_nat: Some(snat_v6),
+                    ephemeral_ip: Some(eip_v6),
+                    floating_ips: fips_v6,
+                }
+                .into())
+            }
+            (Some(snat_v4), None, None) => {
+                if !fips_v6.is_empty() {
+                    return Err(ExternalIpsError::MixedIpVersions);
+                }
+                Ok(ExternalIpv4Config {
+                    source_nat: Some(snat_v4),
+                    ephemeral_ip: None,
+                    floating_ips: fips_v4,
+                }
+                .into())
+            }
+            (Some(snat_v4), None, Some(IpAddr::V4(eip_v4))) => {
+                if !fips_v6.is_empty() {
+                    return Err(ExternalIpsError::MixedIpVersions);
+                }
+                Ok(ExternalIpv4Config {
+                    source_nat: Some(snat_v4),
+                    ephemeral_ip: Some(eip_v4),
+                    floating_ips: fips_v4,
+                }
+                .into())
+            }
+            (Some(_), None, Some(IpAddr::V6(_)))
+            | (None, Some(_), Some(IpAddr::V4(_))) => {
+                Err(ExternalIpsError::MixedIpVersions)
+            }
+            (Some(_), Some(_), None) | (Some(_), Some(_), Some(_)) => {
+                Err(ExternalIpsError::MixedIpVersions)
+            }
+        }
+    }
+}
+
+#[derive(Debug, thiserror::Error)]
+enum ExternalIpsError {
+    #[error(
+        "Must specify at least one SNAT, ephemeral, or floating IP address"
+    )]
+    NoIps,
+    #[error(
+        "SNAT, ephemeral, and floating IPs must all be of the same IP version"
+    )]
+    MixedIpVersions,
+    #[error(transparent)]
+    SourceNat(#[from] SourceNatConfigError),
 }
