+++ sled-agent/types/versions/src/add_dual_stack_shared_network_interfaces/instance.rs
@@ -11,7 +11,6 @@ use omicron_common::api::internal::nexus::HostIdentifier;
 use omicron_common::api::internal::nexus::VmmRuntimeState;
 use omicron_common::api::internal::shared::DhcpConfig;
 use omicron_common::api::internal::shared::NetworkInterface;
-use omicron_common::api::internal::shared::external_ip::v1::SourceNatConfig;
 use omicron_uuid_kinds::InstanceUuid;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -19,6 +18,7 @@ use uuid::Uuid;
 
 use crate::v1::instance::InstanceMetadata;
 use crate::v1::instance::VmmSpec;
+use crate::v1::inventory::SourceNatConfig;
 use crate::v7::instance::InstanceMulticastMembership;
 use crate::v9::instance::DelegatedZvol;
