+++ sled-agent/types/versions/src/add_icmpv6_firewall_support/instance.rs
@@ -14,7 +14,6 @@ use omicron_common::api::internal::nexus::VmmRuntimeState;
 use omicron_common::api::internal::shared::DelegatedZvol;
 use omicron_common::api::internal::shared::DhcpConfig;
 use omicron_common::api::internal::shared::NetworkInterface;
-use omicron_common::api::internal::shared::external_ip::v1::ExternalIpConfig;
 use omicron_uuid_kinds::InstanceUuid;
 use schemars::JsonSchema;
 use serde::Deserialize;
@@ -24,6 +23,7 @@ use uuid::Uuid;
 use crate::v1::instance::InstanceMetadata;
 use crate::v7::instance::InstanceMulticastMembership;
 use crate::v10;
+use crate::v11::instance::ExternalIpConfig;
 use crate::v18;
 use crate::v18::attached_subnet::AttachedSubnet;
 use crate::v29;
