+++ sled-agent/types/versions/src/delegate_zvol_to_propolis/instance.rs
@@ -7,7 +7,6 @@ use std::net::{IpAddr, SocketAddr};
 use omicron_common::api::external::Hostname;
 use omicron_common::api::internal::nexus::VmmRuntimeState;
 use omicron_common::api::internal::shared::DhcpConfig;
-use omicron_common::api::internal::shared::external_ip::v1::SourceNatConfig;
 use omicron_common::api::internal::shared::network_interface::v1::NetworkInterface;
 use omicron_uuid_kinds::DatasetUuid;
 use omicron_uuid_kinds::ExternalZpoolUuid;
@@ -19,6 +18,7 @@ use uuid::Uuid;
 use crate::v1::instance::InstanceMetadata;
 use crate::v1::instance::ResolvedVpcFirewallRule;
 use crate::v1::instance::VmmSpec;
+use crate::v1::inventory::SourceNatConfig;
 use crate::v7;
 use crate::v7::instance::InstanceMulticastMembership;
