+++ sled-agent/types/versions/src/initial/instance.rs
@@ -11,13 +11,14 @@ use omicron_common::api::external;
 use omicron_common::api::external::Hostname;
 use omicron_common::api::internal::nexus::{HostIdentifier, VmmRuntimeState};
 use omicron_common::api::internal::shared::DhcpConfig;
-use omicron_common::api::internal::shared::external_ip::v1::SourceNatConfig;
 use omicron_common::api::internal::shared::network_interface::v1::NetworkInterface;
 use omicron_uuid_kinds::{InstanceUuid, PropolisUuid};
 use propolis_api_types_versions::v1::instance_spec::InstanceSpec;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
+
+use super::inventory::SourceNatConfig;
 
 /// Path parameters for VMM requests.
 #[derive(Deserialize, JsonSchema)]
