+++ sled-agent/types/versions/src/make_all_external_ip_fields_optional/instance.rs
@@ -4,13 +4,16 @@
 
 //! Instance types for version `MAKE_ALL_EXTERNAL_IP_FIELDS_OPTIONAL`.
 
+use std::collections::BTreeSet;
+use std::net::Ipv4Addr;
+use std::net::Ipv6Addr;
 use std::net::SocketAddr;
 
+use omicron_common::address::ConcreteIp;
 use omicron_common::api::external::Hostname;
 use omicron_common::api::internal::nexus::VmmRuntimeState;
 use omicron_common::api::internal::shared::DelegatedZvol;
 use omicron_common::api::internal::shared::DhcpConfig;
-use omicron_common::api::internal::shared::ExternalIpConfig;
 use omicron_common::api::internal::shared::NetworkInterface;
 use omicron_uuid_kinds::InstanceUuid;
 use schemars::JsonSchema;
@@ -20,6 +23,9 @@ use uuid::Uuid;
 
 use crate::v1::instance::InstanceMetadata;
 use crate::v7::instance::InstanceMulticastMembership;
+use crate::v11;
+use crate::v11::inventory::SnatSchema;
+use crate::v11::inventory::SourceNatConfig;
 use crate::v18::attached_subnet::AttachedSubnet;
 use crate::v29::instance::VmmSpec;
 use crate::v31;
@@ -98,6 +104,147 @@ impl From<v31::instance::InstanceSledLocalConfig> for InstanceSledLocalConfig {
             firewall_rules: old.firewall_rules,
             dhcp_config: old.dhcp_config,
             delegated_zvols: old.delegated_zvols,
+        }
+    }
+}
+
+/// Helper trait specifying the name of the JSON Schema for an
+/// `ExternalIpConfig` object.
+///
+/// This exists so we can use a generic type and have the names of the concrete
+/// type aliases be the same as the name of the schema oject.
+pub trait ExternalIpSchema {
+    fn json_schema_name() -> String;
+}
+
+impl ExternalIpSchema for Ipv4Addr {
+    fn json_schema_name() -> String {
+        String::from("ExternalIpv4Config")
+    }
+}
+
+impl ExternalIpSchema for Ipv6Addr {
+    fn json_schema_name() -> String {
+        String::from("ExternalIpv6Config")
+    }
+}
+
+/// External IP address configuration.
+///
+/// This encapsulates all the external addresses of a single IP version,
+/// including source NAT, Ephemeral, and Floating IPs.
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
+#[serde(bound = "T: ConcreteIp + SnatSchema + serde::de::DeserializeOwned")]
+pub struct ExternalIps<T>
+where
+    T: ConcreteIp,
+{
+    /// Source NAT configuration, for outbound-only connectivity.
+    pub source_nat: Option<SourceNatConfig<T>>,
+    /// An Ephemeral address for in- and outbound connectivity.
+    pub ephemeral_ip: Option<T>,
+    /// Additional Floating IPs for in- and outbound connectivity.
+    pub floating_ips: BTreeSet<T>,
+}
+
+impl<T: ConcreteIp> Default for ExternalIps<T> {
+    fn default() -> Self {
+        Self {
+            source_nat: None,
+            ephemeral_ip: None,
+            floating_ips: BTreeSet::new(),
+        }
+    }
+}
+
+pub type ExternalIpv4Config = ExternalIps<Ipv4Addr>;
+pub type ExternalIpv6Config = ExternalIps<Ipv6Addr>;
+
+// Private type only used for deriving the JSON schema for `ExternalIps`.
+#[derive(JsonSchema)]
+#[allow(dead_code)]
+struct ExternalIpsSchema<T>
+where
+    T: ConcreteIp + SnatSchema + ExternalIpSchema,
+{
+    /// Source NAT configuration, for outbound-only connectivity.
+    source_nat: Option<SourceNatConfig<T>>,
+    /// An Ephemeral address for in- and outbound connectivity.
+    ephemeral_ip: Option<T>,
+    /// Additional Floating IPs for in- and outbound connectivity.
+    floating_ips: BTreeSet<T>,
+}
+
+impl JsonSchema for ExternalIpv4Config {
+    fn schema_name() -> String {
+        String::from("ExternalIpv4Config")
+    }
+
+    fn json_schema(
+        generator: &mut schemars::r#gen::SchemaGenerator,
+    ) -> schemars::schema::Schema {
+        ExternalIpsSchema::<Ipv4Addr>::json_schema(generator)
+    }
+}
+
+impl JsonSchema for ExternalIpv6Config {
+    fn schema_name() -> String {
+        String::from("ExternalIpv6Config")
+    }
+
+    fn json_schema(
+        generator: &mut schemars::r#gen::SchemaGenerator,
+    ) -> schemars::schema::Schema {
+        ExternalIpsSchema::<Ipv6Addr>::json_schema(generator)
+    }
+}
+
+/// External IP address configuration.
+///
+/// This encapsulates all external addresses for an instance, with separate
+/// optional configurations for IPv4 and IPv6.
+#[derive(
+    Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize,
+)]
+pub struct ExternalIpConfig {
+    /// IPv4 external IP configuration.
+    pub v4: Option<ExternalIpv4Config>,
+    /// IPv6 external IP configuration.
+    pub v6: Option<ExternalIpv6Config>,
+}
+
+impl From<v11::instance::ExternalIpv4Config> for ExternalIpv4Config {
+    fn from(old: v11::instance::ExternalIpv4Config) -> Self {
+        Self {
+            source_nat: old.source_nat,
+            ephemeral_ip: old.ephemeral_ip,
+            floating_ips: old.floating_ips.into_iter().collect(),
+        }
+    }
+}
+
+impl From<v11::instance::ExternalIpv6Config> for ExternalIpv6Config {
+    fn from(old: v11::instance::ExternalIpv6Config) -> Self {
+        Self {
+            source_nat: old.source_nat,
+            ephemeral_ip: old.ephemeral_ip,
+            floating_ips: old.floating_ips.into_iter().collect(),
+        }
+    }
+}
+
+impl From<v11::instance::ExternalIpConfig> for ExternalIpConfig {
+    fn from(old: v11::instance::ExternalIpConfig) -> Self {
+        match old {
+            v11::instance::ExternalIpConfig::V4(v4) => {
+                Self { v4: Some(v4.into()), v6: None }
+            }
+            v11::instance::ExternalIpConfig::V6(v6) => {
+                Self { v4: None, v6: Some(v6.into()) }
+            }
+            v11::instance::ExternalIpConfig::DualStack { v4, v6 } => {
+                Self { v4: Some(v4.into()), v6: Some(v6.into()) }
+            }
         }
     }
 }
