+++ sled-agent/types/versions/src/multicast_support/instance.rs
@@ -7,7 +7,6 @@ use std::net::{IpAddr, SocketAddr};
 use omicron_common::api::external::Hostname;
 use omicron_common::api::internal::nexus::VmmRuntimeState;
 use omicron_common::api::internal::shared::DhcpConfig;
-use omicron_common::api::internal::shared::external_ip::v1::SourceNatConfig;
 use omicron_common::api::internal::shared::network_interface::v1::NetworkInterface;
 use omicron_uuid_kinds::InstanceUuid;
 use schemars::JsonSchema;
@@ -18,6 +17,7 @@ use crate::v1;
 use crate::v1::instance::InstanceMetadata;
 use crate::v1::instance::ResolvedVpcFirewallRule;
 use crate::v1::instance::VmmSpec;
+use crate::v1::inventory::SourceNatConfig;
 
 /// The body of a request to ensure that a instance and VMM are known to a sled
 /// agent.
