+++ sled-agent/types/versions/src/two_types_of_delegated_zvol/instance.rs
@@ -9,7 +9,6 @@ use omicron_common::api::internal::nexus::VmmRuntimeState;
 use omicron_common::api::internal::shared::DelegatedZvol;
 use omicron_common::api::internal::shared::DhcpConfig;
 use omicron_common::api::internal::shared::NetworkInterface;
-use omicron_common::api::internal::shared::external_ip::v1::ExternalIpConfig;
 use omicron_uuid_kinds::InstanceUuid;
 
 use schemars::JsonSchema;
@@ -22,6 +21,7 @@ use crate::v7::instance::InstanceMulticastMembership;
 use crate::v9;
 use crate::v10::instance::ResolvedVpcFirewallRule;
 use crate::v11;
+use crate::v11::instance::ExternalIpConfig;
 
 /// The body of a request to ensure that a instance and VMM are known to a sled
 /// agent.
